+++ packages/core/src/hooks/data/useCreate.ts
@@ -1,5 +1,6 @@
 import { useMutation, UseMutationResult } from "@tanstack/react-query";
 import pluralize from "pluralize";
+import { pickDataProvider } from "@definitions/helpers";
 
 import {
     CreateResponse,
@@ -10,6 +11,7 @@ import {
     IQueryKeys,
 } from "../../interfaces";
 import {
+    useResource,
     useTranslate,
     useCheckError,
     usePublish,
@@ -75,6 +77,8 @@ export const useCreate = <
     const dataProvider = useDataProvider();
     const invalidateStore = useInvalidate();
 
+    const { resources } = useResource();
+
     const translate = useTranslate();
     const publish = usePublish();
     const { log } = useLog();
@@ -92,7 +96,9 @@ export const useCreate = <
             metaData,
             dataProviderName,
         }: useCreateParams<TVariables>) => {
-            return dataProvider(dataProviderName).create<TData, TVariables>({
+            return dataProvider(
+                pickDataProvider(resource, dataProviderName, resources),
+            ).create<TData, TVariables>({
                 resource,
                 variables: values,
                 metaData,
@@ -135,7 +141,11 @@ export const useCreate = <
 
                 invalidateStore({
                     resource,
-                    dataProviderName,
+                    dataProviderName: pickDataProvider(
+                        resource,
+                        dataProviderName,
+                        resources,
+                    ),
                     invalidates,
                 });
 
@@ -156,7 +166,11 @@ export const useCreate = <
                     resource,
                     data: values,
                     meta: {
-                        dataProviderName,
+                        dataProviderName: pickDataProvider(
+                            resource,
+                            dataProviderName,
+                            resources,
+                        ),
                         id: data?.data?.id ?? undefined,
                         ...rest,
                     },